Move ThreadLocalCleaner to parent package and make it package protected
